src/requirements.txt
fastapi[all]
uvicorn
pymongo
pydantic
bcrypt
pyjwt
